+++ src/display/Graphics.js
@@ -11,15 +11,31 @@ class Graphics extends DisplayObject {
    * Creates new Graphics instance.
    *
    * @param {GraphicsData|string|null} graphicsData The id of BVG object.
+   * @param {boolean} trim Flag to determine the passed graphicsData needs trim.
    */
-  constructor(graphicsData = null) {
+  constructor(graphicsData = null, trim = false) {
     super();
 
     /** @private @type {Rectangle} */
     this.mBounds = new Rectangle();
 
-    /** @private @type {GraphicsData} */
-    this.mGraphicsData;
+    /**
+     * For internal usage
+     *
+     * @private @type {Rectangle|null} */
+    this.mLocalBounds = null;
+
+    /** @private @type {GraphicsData|null} */
+    this.mGraphicsData = null;
+
+    /** @private @type {number} */
+    this.mDataOffsetX = 0;
+
+    /** @private @type {number} */
+    this.mDataOffsetY = 0;
+
+    /** @private @type {boolean} */
+    this.mTrim = trim;
 
     if (graphicsData === null) {
       this.mGraphicsData = new GraphicsData();
@@ -27,6 +43,16 @@ class Graphics extends DisplayObject {
       this.mGraphicsData = AssetManager.default.getGraphicsData(graphicsData);
     } else {
       this.mGraphicsData = graphicsData;
+    }
+
+    if (trim) {
+      this.mGraphicsData.onGetLocalBounds(this, new Matrix());
+
+      if (this.mLocalBounds) {
+        this.mDataOffsetX = this.mLocalBounds.x;
+        this.mDataOffsetY = this.mLocalBounds.y;
+        this.mLocalBounds = null;
+      }
     }
   }
 
@@ -48,7 +74,19 @@ class Graphics extends DisplayObject {
       return outRect;
     }
 
-    return this.mGraphicsData.onGetLocalBounds(outRect, new Matrix());
+    this.mGraphicsData.onGetLocalBounds(this, new Matrix());
+
+    this.mLocalBounds && outRect.copyFrom(this.mLocalBounds);
+    this.mLocalBounds = null;
+
+    if (!this.mTrim) {
+      outRect.width += Math.max(0, outRect.x);
+      outRect.height += Math.max(0, outRect.y);
+      outRect.x = Math.min(0, outRect.x);
+      outRect.y = Math.min(0, outRect.y);
+    }
+
+    return outRect;
   }
 
   /**
@@ -77,6 +115,30 @@ class Graphics extends DisplayObject {
    */
   fillStyle(color = 0, alpha = 1) {
     this.mGraphicsData.fillStyle(color, alpha);
+  }
+
+  /**
+   * Sets fill style to gradient.
+   *
+   * @public
+   * @param {GraphicsGradient} gradient Fill gradient.
+   *
+   * @returns {void}
+   */
+  fillGradient(gradient) {
+    this.mGraphicsData.fillGradient(gradient);
+  }
+
+  /**
+   * Sets fill style to pattern.
+   *
+   * @public
+   * @param {GraphicsPattern} pattern Fill pattern.
+   *
+   * @returns {void}
+   */
+  fillPattern(pattern) {
+    this.mGraphicsData.fillPattern(pattern);
   }
 
   /**
@@ -160,6 +222,22 @@ class Graphics extends DisplayObject {
   }
 
   /**
+   * Creates closed rounded rectangle.
+   *
+   * @public
+   * @param {number} x
+   * @param {number} y
+   * @param {number} width
+   * @param {number} height
+   * @param {number} radius
+   *
+   * @returns {void}
+   */
+  roundedRect(x, y, width, height, radius) {
+    this.mGraphicsData.roundedRect(x, y, width, height, radius);
+  }
+
+  /**
    * @public
    * @param {number} cp1x
    * @param {number} cp1y
@@ -239,23 +317,8 @@ class Graphics extends DisplayObject {
     this.mGraphicsData.fill(isNonZero);
     this.setTransformDirty();
   }
-}
 
-class GraphicsPath {
-  constructor() {
-    /** @type {Rectangle|null} */
-    this.bounds = null;
-
-    /** @type {Array<number>} */
-    this.points = [];
-
-    /** @type {number} */
-    this.maxLineWidth = 0;
-
-    /** @type {number} */
-    this.lastLineWidth = 0;
-
-    /** @type {number} */
-    this.lineMul = 0.5;
+  createLinearGradient(x, y, width, height) {
+    return new GraphicsLinearGradient(x, y, width, height);
   }
 }