+++ src/display/Stage.js
@@ -30,8 +30,12 @@ class Stage extends GameObject {
 
     /** @private @type {number} */
     this.mCacheWidth = 0;
+    
     /** @private @type {number} */
     this.mCacheHeight = 0;
+
+    /** @private @type {number} */
+    this.mDPR = Device.getDevicePixelRatio();
 
     /** @private @type {StageOrientation} */
     this.mOrientation = StageOrientation.UNIVERSAL;
@@ -100,7 +104,7 @@ class Stage extends GameObject {
   __refresh() {
     let size = Black.instance.viewport.size.clone();
 
-    if (this.mOrientationLock && this.mOrientation === StageOrientation.LANDSCAPE && Device.isPortrait)
+    if (this.mOrientationLock === true && this.mOrientation === StageOrientation.LANDSCAPE && Device.isPortrait)
       [size.width, size.height] = [size.height, size.width];
 
     let windowWidth = size.width;
@@ -137,10 +141,10 @@ class Stage extends GameObject {
 
       this.mScaleX = this.mScaleY = this.mStageScaleFactor = Math.min(windowWidth / width, windowHeight / height);
     } else if (this.mScaleMode === StageScaleMode.NO_SCALE) {
-      this.mStageWidth = (size.width * this.dpr);
-      this.mStageHeight = (size.height * this.dpr);
+      this.mStageWidth = (size.width * this.mDPR);
+      this.mStageHeight = (size.height * this.mDPR);
 
-      this.mScaleX = this.mScaleY = this.mStageScaleFactor = 1 / this.dpr;
+      this.mScaleX = this.mScaleY = this.mStageScaleFactor = 1 / this.mDPR;
     } else {
       Debug.error('Not supported stage scale mode.');
     }
@@ -202,17 +206,6 @@ class Stage extends GameObject {
   }
 
   /**
-   * Device Pixel Ratio.
-   * 
-   * @public
-   * @readonly
-   * @returns {number}
-   */
-  get dpr() {
-    return Device.getDevicePixelRatio();
-  }
-
-  /**
    * Stage scale factor.
    * 
    * @public
@@ -231,7 +224,7 @@ class Stage extends GameObject {
    * @returns {number}
    */
   get renderWidth() {
-    return this.mStageWidth * this.dpr * this.mStageScaleFactor;
+    return this.mStageWidth * this.mDPR * this.mStageScaleFactor;
   }
 
   /**
@@ -242,7 +235,7 @@ class Stage extends GameObject {
    * @returns {number}
    */
   get renderHeight() {
-    return this.mStageHeight * this.dpr * this.mStageScaleFactor;
+    return this.mStageHeight * this.mDPR * this.mStageScaleFactor;
   }
 
   /**
@@ -282,6 +275,7 @@ class Stage extends GameObject {
    */
   set orientationLock(value) {
     this.mOrientationLock = value;
+    this.__refresh();
   }
 
   /**
@@ -306,6 +300,9 @@ class Stage extends GameObject {
   get localTransformation() {
     this.mLocalTransform.set(this.mScaleX, 0, 0, this.mScaleY, this.mX, this.mY);
 
+    if (this.mOrientationLock === false)
+      return this.mLocalTransform;
+
     // orientation lock hacks
     // chrome   window.screen.orientation.type
     // firefox  window.screen.mozOrientation
@@ -329,12 +326,16 @@ class Stage extends GameObject {
     }
 
     if (this.mOrientation === StageOrientation.LANDSCAPE && Device.isPortrait) {
+
       this.mLocalTransform.rotate((angle + 90) * Math.PI / 180);
+
       let x = (Black.instance.viewport.size.width * 0.5) + (this.mStageHeight * 0.5 * this.mStageScaleFactor);
       let y = (Black.instance.viewport.size.height * 0.5) - (this.mStageWidth * 0.5 * this.mStageScaleFactor);
 
       this.mLocalTransform.setTranslation(x, y);
+
     } else if (this.mOrientation === StageOrientation.PORTRAIT && Device.isLandscape) {
+
       this.mLocalTransform.rotate((angle + 180) * Math.PI / 180);
 
       let x = (Black.instance.viewport.size.width * 0.5) - (this.mStageHeight * 0.5 * this.mStageScaleFactor);