+++ src/drivers/canvas/DisplayObjectRendererCanvas.js
@@ -4,6 +4,7 @@
  * @extends Renderer
  * @cat drivers.canvas
  */
+
 /* @echo EXPORT */
 class DisplayObjectRendererCanvas extends Renderer {
   constructor() {
@@ -23,29 +24,47 @@ class DisplayObjectRendererCanvas extends Renderer {
 
     /** @private @type {boolean} */
     this.mIsClipped = false;
+
+    this.mIsCached = false;
+
+    /** @private @type {Matrix|null} */
+    this.mBakeInvertedMatrix = null;
   }
 
   /** @inheritDoc */
   preRender(driver, session) {
     let gameObject = /** @type {DisplayObject} */ (this.gameObject);
 
-    this.mIsClipped = gameObject.mClipRect !== null && gameObject.mClipRect.isEmpty === false;
+    this.mIsClipped = gameObject.mClipRect !== null;
     this.endPassRequired = this.mIsClipped;
 
     if (gameObject.mCacheAsBitmap === true) {
-      let isStatic = gameObject.checkStatic(true);
-      if (isStatic === true && this.mCacheAsBitmapDirty === true) {
-        gameObject.setTransformDirty();
-        this.__refreshBitmapCache();
-        this.mCacheAsBitmapDirty = false;
-        this.endPassRequired = false;
-      } else if (isStatic === false) {
-        this.mCacheAsBitmapDirty = true;
-        gameObject.mDirty |= DirtyFlag.RENDER;
+      if (gameObject.mCacheAsBitmapDynamic === false) {
+        if (this.mIsCached === false) {
+          this.mIsCached = true;
+          gameObject.setTransformDirty();
+          this.__refreshBitmapCache();
+          this.mCacheAsBitmapDirty = false;
+          this.endPassRequired = false;
+        } else {
+          gameObject.mDirty |= DirtyFlag.RENDER;
+        }
+      } else {
+        let isStatic = gameObject.checkStatic(true);
+
+        if (isStatic === true && this.mCacheAsBitmapDirty === true) {
+          gameObject.setTransformDirty();
+          this.__refreshBitmapCache();
+          this.mCacheAsBitmapDirty = false;
+          this.endPassRequired = false;
+        } else if (isStatic === false) {
+          this.mCacheAsBitmapDirty = true;
+          gameObject.mDirty |= DirtyFlag.RENDER;
+        }
       }
     }
 
-    this.skipChildren = gameObject.mCacheAsBitmap === true && this.mCacheAsBitmapDirty === false;
+    this.skipChildren = gameObject.mCacheAsBitmap === true && this.mCacheAsBitmapDirty === false || this.mIsClipped && gameObject.mClipRect.isEmpty;
     this.skipSelf = false;
 
     if (this.skipChildren === true) {
@@ -61,7 +80,7 @@ class DisplayObjectRendererCanvas extends Renderer {
       this.alpha = 1;
       this.blendMode = BlendMode.NORMAL;
       this.color = null;
-      this.skipSelf = gameObject.mAlpha <= 0 || gameObject.mVisible === false;
+      this.skipSelf = gameObject.mAlpha <= 0 || gameObject.mVisible === false || this.mIsClipped && gameObject.mClipRect.isEmpty;
     }
     else {
       this.alpha = gameObject.mAlpha * this.parent.alpha;
@@ -69,7 +88,7 @@ class DisplayObjectRendererCanvas extends Renderer {
       this.blendMode = gameObject.mBlendMode === BlendMode.AUTO ? this.parent.blendMode : gameObject.mBlendMode;
 
       this.skipChildren = gameObject.mAlpha <= 0 || gameObject.mVisible === false;
-      this.skipSelf = this.skipChildren === true || this.mIsClipped === false;
+      this.skipSelf = this.skipChildren === true || this.mIsClipped === false || this.mIsClipped && gameObject.mClipRect.isEmpty;
     }
   }
 
@@ -78,8 +97,16 @@ class DisplayObjectRendererCanvas extends Renderer {
     let gameObject = /** @type {DisplayObject} */ (this.gameObject);
     let transform = gameObject.worldTransformation;
 
-    if (this.skipChildren === true)
+    if (this.skipChildren === true && this.mCacheAsBitmapMatrixCache) {
       transform = this.mCacheAsBitmapMatrixCache;
+
+      if (gameObject.mCacheAsBitmapDynamic === false) {
+        transform = new Matrix()
+          .append(this.gameObject.worldTransformation)
+          .append(this.mBakeInvertedMatrix)
+          .append(this.mCacheAsBitmapMatrixCache);
+      }
+    }
 
     if (this.skipChildren === true || this.endPassRequired === true) {
       driver.setSnapToPixels(gameObject.snapToPixels);
@@ -101,7 +128,7 @@ class DisplayObjectRendererCanvas extends Renderer {
   __refreshBitmapCache() {
     const bounds = this.gameObject.getBounds(Black.stage, true);
     const sf = Black.stage.scaleFactor;
-    const fs = Black.driver.finalScale;
+    const fs = Black.driver.renderScaleFactor * sf;
 
     /** @type {Matrix} */
     let m = Matrix.pool.get();
@@ -135,6 +162,7 @@ class DisplayObjectRendererCanvas extends Renderer {
     this.mCacheAsBitmapMatrixCache.data[4] = -this.mCacheAsBitmapMatrixCache.data[4];
     this.mCacheAsBitmapMatrixCache.data[5] = -this.mCacheAsBitmapMatrixCache.data[5];
 
+    this.mBakeInvertedMatrix = this.gameObject.worldTransformationInverted.clone();
     //this.mCacheTexture.__dumpToDocument();
   }
 }