+++ CHANGELOG.md
@@ -1,6 +1,52 @@
 Black Engine: Changelog
 =======================
 
+# v0.5.8
+### New features
+- Added `useHiDPR` property to control render quality on high DPI screens
+- Added ability to draw rounded rectangle
+- Added ability to load custom asset in AssetManager using CustomAsset
+- Added gradient and pattern support for Graphics
+
+### Changes
+- Changed empty clip rectangle to clip everything instead of nothing
+- Improved anchor performance
+
+### Bug Fixes
+- Fixed Orientation Lock still works when disabled
+- Fixed many GCC comments
+- Fixed bug causing colored sprite to render incorrectly
+- Fixed slice9grid not working as expected on colored sprites
+- Fixed bug causing Arcade physics to not cleanup colliders after removing objects from stage
+- Fixed Arcade physics units to be always within stage
+- Fixed clip rect not working on TextField
+- Fixed `getTextures` returning null for backed vectors
+- Fixed BVG quadratic curve math
+- Fixed many BVG rendering bugs and glitches
+- Fixed anchors not working in some cases
+
+# v0.5.7
+### Changes
+- Changed license to Simplified BSD!!!
+- Changed `Emitter#state` to be public
+- Added `TextField#getAllStyles` method which returns all styles 
+
+### Bug Fixes
+- Fixed `GameObject#relativeTo` method
+- Fixed tiling sprite causing incorrect rendering of other Graphics objects
+
+# v0.5.6
+### New features
+- Added `Graphics#cacheAsBitmapDynamic` property allowing to disable auto refresh of bitmap cache
+
+### Bug Fixes
+- Fixed few Cache As Bitmap bugs causing Graphics and tiling sprite to render incorrectly
+- Fixed incorrect rendering of cached slice9grid texture
+- Fixed incorrect rendering of tiled texture from texture atlas
+
+### Examples & Docs
+- Added "Cache As Bitmap Dynamic" example
+
 # v0.5.5
 ### New features
 - Added support for Black Vector Graphics (beta)